+++ ViewModel/HospitalViewModel.cs
@@ -26,7 +26,7 @@ namespace Modern_Real_Estate.ViewModel
         public string TextBoxValueCountry { get; set; }
         public int TextBoxValueArea { get; set; }
         public string TextBoxValueType { get; }
-        public double TextBoxValuePrice { get; set; }
+        public decimal TextBoxValuePrice { get; set; }
 
         public RelayCommand AddCommand => new RelayCommand(execute => AddEstate());
         public RelayCommand DeleteCommand => new RelayCommand(execute => DeleteEstate(), canExecute => SelectedEstate != null);
